# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Public read-only API for portfolio data",
        "title": "Portfolio Public API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
...
        "/miniatures": {
            "get": {
                "description": "Get list of all miniature painting projects with images",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "miniatures"
                ],
                "summary": "Get all miniature projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_GunarsK-portfolio_public-api_internal_models.MiniatureProject"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/miniatures/themes": {
            "get": {
                "description": "Get list of all miniature themes with their associated projects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "miniatures"
                ],
                "summary": "Get all miniature themes with projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_GunarsK-portfolio_public-api_internal_models.MiniatureTheme"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/miniatures/{id}": {
            "get": {
                "description": "Get detailed information about a specific miniature project",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "miniatures"
                ],
                "summary": "Get miniature project by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Miniature Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_GunarsK-portfolio_public-api_internal_models.MiniatureProject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_GunarsK-portfolio_portfolio-common_models.Image": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_GunarsK-portfolio_portfolio-common_models.MiniaturePaint": {
            "type": "object",
            "required": [
                "manufacturer",
                "name"
            ],
            "properties": {
                "colorHex": {
                    "description": "ColorHex is the hexadecimal color code in #RRGGBB or #RGB format (e.g., #FF5733, #F00)",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "manufacturer": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "paintType": {
                    "description": "PaintType categorizes the paint (Base, Layer, Shade, Wash, Contrast, Dry, Technical, Metallic, Air, Primer, Edge, Glaze, Ink)\nDatabase enforces these values via CHECK constraint",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_GunarsK-portfolio_portfolio-common_models.MiniatureProject": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "completedDate": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "string"
                },
                "displayOrder": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.Image"
                    }
                },
                "manufacturer": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "paints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.MiniaturePaint"
                    }
                },
                "scale": {
                    "type": "string"
                },
                "techniques": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "theme": {
                    "description": "Associations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.MiniatureTheme"
                        }
                    ]
                },
                "themeId": {
                    "type": "integer"
                },
                "timeSpent": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_GunarsK-portfolio_portfolio-common_models.MiniatureTheme": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "coverImageFile": {
                    "description": "Associations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.StorageFile"
                        }
                    ]
                },
                "coverImageId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayOrder": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "miniatures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.MiniatureProject"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_GunarsK-portfolio_portfolio-common_models.StorageFile": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer"
                },
                "fileType": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mimeType": {
                    "type": "string"
                },
                "url": {
                    "description": "Computed field",
                    "type": "string"
                }
            }
        },
...
        "github_com_GunarsK-portfolio_public-api_internal_models.MiniatureProject": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "completedDate": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "string"
                },
                "displayOrder": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.Image"
                    }
                },
                "manufacturer": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "paints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.MiniaturePaint"
                    }
                },
                "scale": {
                    "type": "string"
                },
                "techniques": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "theme": {
                    "description": "Associations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.MiniatureTheme"
                        }
                    ]
                },
                "themeId": {
                    "type": "integer"
                },
                "timeSpent": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_GunarsK-portfolio_public-api_internal_models.MiniatureTheme": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "coverImageFile": {
                    "description": "Associations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.StorageFile"
                        }
                    ]
                },
                "coverImageId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayOrder": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "miniatures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.MiniatureProject"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}